requirements.txt
SpeechRecognition
pymongo
pathlib
datetime
kafka-python
elasticsearch==8.15.0
